refactor: 컴포넌트명, 파일명 수정 (#56)
