+++ src/shared/styles/bar-tab.ts
